requirements.txt
asgiref==3.11.0
asttokens==3.0.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Django==5.2.10
djangorestframework==3.16.1
executing==2.2.0
fonttools==4.57.0
geographiclib==2.1
geopy==2.4.1
google-api-core==2.24.2
google-auth==2.40.0
google-cloud-speech==2.32.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==25.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.5
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
pocketsphinx==5.0.4
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
Pygments==2.19.1
pyjokes==0.8.3
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
pywin32==310
pyzmq==26.4.0
requests==2.32.3
rsa==4.9.1
scipy==1.15.2
six==1.17.0
sounddevice==0.5.1
soupsieve==2.7
sqlparse==0.5.5
srt==3.5.3
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.4.0
vosk==0.3.45
wcwidth==0.2.13
websockets==15.0.1
wikipedia==1.4.0
